% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_func.r
\name{as_five_greade_scores}
\alias{as_five_greade_scores}
\title{Convert the percentage system to a five-point system}
\usage{
as_five_greade_scores(data)
}
\arguments{
\item{data}{Input data}
}
\value{
five_greade_scores
}
\description{
Convert the percentage system to a five-point system
}
\examples{
as_five_greade_scores(c(59,65,75,86,99))
}
