% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_func.r
\name{unscale}
\alias{unscale}
\title{Accept the result of scale and perform the inverse transformation}
\usage{
unscale(data_scaled)
}
\arguments{
\item{data_scaled}{the result of scale}
}
\value{
the result of unscale
}
\description{
Accept the result of scale and perform the inverse transformation
}
\examples{
unscale(scale(1:5))
}
