% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_func.r
\name{which_max}
\alias{which_max}
\title{Accept a numeric vector as a parameter and output the index of the maximum value}
\usage{
which_max(data)
}
\arguments{
\item{data}{Input data}
}
\value{
the index of the maximum value
}
\description{
Accept a numeric vector as a parameter and output the index of the maximum value
}
\examples{
which_max(1:100)
}
